sign_up request to db finish
