+++ server/inc/server.h
@@ -27,10 +27,15 @@ typedef struct s_new_client {
 }              t_new_client;
 
 
-
 GHashTable **mx_get_online_users(void);
+sqlite3 **mx_get_db(void); 
+
 GDataOutputStream *mx_get_socket_by_user_id(gint64 user_id);
 gssize mx_send_data(GDataOutputStream *data_out, gchar *data);
 
+//Sign_up_rq
+gint mx_sign_up_quary(cJSON *root, sqlite3 *db);
+gboolean mx_check_if_user_excist(cJSON *root, sqlite3 *db);
+void mx_sign_up(cJSON *root, t_client *client);
 
 #endif /* end of include guard: SERVER_H */
